Rename files and fix includes
